requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
cryptography
python-dotenv
